fix: updating data types to fix hashing scheme
The Example.js provided in the original EIP712 spec uses ethutil.sha3 as it's hashing function.  We're currently using the smaller keccak256 package, which returns strings rather than buffers.  This replaces the string hashes returned by keccak256() with a Buffer constructed from the hash. Also the Type.encode() function previously duplicated the first element in the concatenated array of abi encoded struct memebers; this fixes that as well.