feat: add topological sort to Domain.fromSignatureRequest
Ensure that when creating a domain from a signature request, that the structure type dependency graph is acyclic, and that the types are instantiated in the proper order.
Additionally add new function isNotStructureType to primitives.js, to detect whether a particular type name is a primitive type or array of primitive types.